Execute clang-tidy as a gate, and fix what it found at the cause
The verifier now runs clang-tidy over every first-party unit, each from the
compile database of the build that compiles it, and refuses a missing
database or a check group the installed clang-tidy does not run.

Its first run found 83 findings. Enums take one byte, widened arithmetic
happens in size_t, title IDs parse with from_chars, float bit comparison has
one owner (sameBits) instead of scattered memcmp and a duplicate span compare,
parameters that could be swapped become named structs, stop tokens pass by
reference, a capture request consumes its callback, optionals are checked,
and the draw-tree tests take their seed from the harness (--seed N). Two
signatures the fork's upstream interface imposes carry a named NOLINT.

+++ src/wiiuport/interp/ContinuousInterpolator.h
@@ -54,7 +54,7 @@ class ContinuousInterpolator final : public frame::FrameEndListener {
     // leads into.
     static constexpr float kBlendPoint = 0.5f;
 
-    enum class Skip : uint32_t {
+    enum class Skip : uint8_t {
         Disabled,
         // A one-shot replay or null diff owns the next frame boundary.
         OneShotInFlight,
@@ -143,7 +143,7 @@ class ContinuousInterpolator final : public frame::FrameEndListener {
     // Where an interpolated tick's time goes, summed over every interpolated
     // tick. The title's own frame waits on all of it, so a tick rate that
     // collapses is attributed to a phase here rather than guessed at.
-    enum class Phase : uint32_t {
+    enum class Phase : uint8_t {
         BlendedReplay,
         Present,
         Restore,
